requirements.txt
Flask==3.0.0
werkzeug==3.0.1
Pillow==10.1.0
numpy==1.24.3
transformers==4.36.0
torch==2.1.0
